Implement a prototype system call
In order to make the context switch work, we set a static stack address
to start at; this address was originally in low memory, but that does
not work on this hardware (data abort), so we tried setting to be a
few pages after the kernel image. This seems to be working for now.

The system call number is set in main.c right before the svc call which
triggers the system from user mode. All state seems to be restored
properly, so I think this is working well enough to break out into a
proper kernel and user land.